input: tri-state-key: Properly filter out spurious interrupts
All of the three pins used for the tri-state key are falling-edge
sensitive, and their IRQs are properly configured as such; however,
spurious interrupts can easily be reproduced where all of the three pins
are in a high state, causing unwanted notification-mode changes as a result.

Add a filter to IRQ worker in order to block the spurious interrupts.

Change-Id: I7fadce119a0ca039eeb563a389ac6aa181f658a3
Signed-off-by: Sultanxda <sultanxda@gmail.com>